+++ src/textual/binding.py
@@ -81,6 +81,8 @@ class Binding:
     If specified in the App's keymap then Textual will use this ID to lookup the binding,
     and substitute the `key` property of the Binding with the key specified in the keymap.
     """
+    system: bool = False
+    """Make this binding a system binding, which removes it from the key panel."""
 
     def parse_key(self) -> tuple[list[str], str]:
         """Parse a key in to a list of modifiers, and the actual key.
@@ -148,6 +150,7 @@ class Binding:
                     priority=binding.priority,
                     tooltip=binding.tooltip,
                     id=binding.id,
+                    system=binding.system,
                 )